IT_Project/pyproject.toml
[tool.poetry]
name = "it-project"
version = "0.1.0"
description = ""
authors = ["eternazzz <alexkonder@mail.ru>"]
readme = "README.md"
packages = [{include = "it_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
